added isSearchAndDropDownMissmatched, renamed Tree to TreeContainer
